backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
duckdb==0.10.3
plotly==5.22.0
scikit-learn==1.5.0
scipy==1.13.1
langchain==0.2.5
langchain-core==0.2.9
langchain-groq==0.1.6
langgraph==0.1.9
langsmith==0.1.77
supabase==2.4.6
httpx==0.27.0
pydantic==2.7.4
openpyxl==3.1.4
pyarrow==16.1.0
